각 버튼 이미지들을 자식으로 넘겨줌
