+++ src/components/common/content/MyInfoItemContent.tsx
@@ -1,7 +1,5 @@
 import styled from '@emotion/styled';
-import AccountIconSvg from '@resources/svg/AccountIconSvg';
-import DeleteUserSvg from '@resources/svg/DeleteUserSvg';
-import SettingSvg from '@resources/svg/SettingIconSvg';
+import { ReactNode } from 'react';
 
 const ItemContainer = styled.div`
   width: 160px;
@@ -18,45 +16,6 @@ const ButtonContainer = styled.div`
   height: 70px;
 `;
 
-const SettingButton = styled(SettingSvg)`
-  cursor: pointer;
-  position: absolute;
-  width: 70px;
-  height: 70px;
-  right: 44px;
-  left: 44px;
-  top: 12px;
-  transition: transform 0.1s ease;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-const AccountButton = styled(AccountIconSvg)`
-  cursor: pointer;
-  position: absolute;
-  width: 70px;
-  height: 70px;
-  right: 44px;
-  left: 44px;
-  top: 12px;
-  transition: transform 0.1s ease;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-const DeleteUserButton = styled(DeleteUserSvg)`
-  cursor: pointer;
-  position: absolute;
-  width: 70px;
-  height: 70px;
-  right: 44px;
-  left: 44px;
-  top: 12px;
-  transition: transform 0.1s ease;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
 const ItmeLabel = styled.div`
   width: 156px;
   height: 43px;
@@ -67,28 +26,13 @@ const ItmeLabel = styled.div`
 
 interface MyInfoContentItmeProps {
   label: string;
-  onClickFunc: (url: string) => void;
-  iconString: string;
+  children: ReactNode;
 }
 
-function MyInfoItemContent({ label, onClickFunc, iconString }: MyInfoContentItmeProps) {
-  let buttonComponent;
-  if (iconString === 'setting') {
-    buttonComponent = <SettingButton onClick={() => onClickFunc('/setting')} />;
-  } else if (iconString === 'account') {
-    buttonComponent = <AccountButton onClick={() => onClickFunc('/register-account')} />;
-  } else {
-    buttonComponent = (
-      <DeleteUserButton
-        onClick={() => {
-          onClickFunc('delete-user');
-        }}
-      />
-    );
-  }
+function MyInfoItemContent({ label, children }: MyInfoContentItmeProps) {
   return (
     <ItemContainer>
-      <ButtonContainer>{buttonComponent}</ButtonContainer>
+      <ButtonContainer>{children}</ButtonContainer>
       <ItmeLabel>{label}</ItmeLabel>
     </ItemContainer>
   );
